+++ packages/sdk/src/types/BlockSpecialEvents.ts
@@ -1,4 +1,6 @@
-import { Amount, BakerId, Base58String } from '../types.js';
+import type { BakerId } from '../types.js';
+import * as CcdAmount from './CcdAmount.js';
+import * as AccountAddress from './AccountAddress.js';
 
 export type BlockSpecialEvent =
     | BlockSpecialEventBakingRewards
@@ -15,19 +17,19 @@ export interface BlockSpecialEventBakingRewards {
     // The amount awarded to each baker.
     bakingRewards: BlockSpecialEventAccountAmount[];
     // The remaining balance of the baker reward account.
-    remainder: Amount;
+    remainder: CcdAmount.Type;
 }
 
 export interface BlockSpecialEventMint {
     tag: 'mint';
     // The amount allocated to the banking reward account.
-    mintBakingReward: Amount;
+    mintBakingReward: CcdAmount.Type;
     // The amount allocated to the finalization reward account.
-    mintFinalizationReward: Amount;
+    mintFinalizationReward: CcdAmount.Type;
     // The amount allocated as the platform development charge.
-    mintPlatformDevelopmentCharge: Amount;
+    mintPlatformDevelopmentCharge: CcdAmount.Type;
     // The account to which the platform development charge is paid.
-    foundationAccount: Base58String;
+    foundationAccount: AccountAddress.Type;
 }
 
 export interface BlockSpecialEventFinalizationRewards {
@@ -35,61 +37,61 @@ export interface BlockSpecialEventFinalizationRewards {
     // The amount awarded to each finalizer.
     finalizationRewards?: BlockSpecialEventAccountAmount[];
     // The remaining balance of the finalization reward account.
-    remainder?: Amount;
+    remainder?: CcdAmount.Type;
 }
 
 export interface BlockSpecialEventBlockReward {
     tag: 'blockReward';
     // The total fees paid for transactions in the block.
-    transactionFees: Amount;
+    transactionFees: CcdAmount.Type;
     // The old balance of the GAS account.
-    oldGasAccount: Amount;
+    oldGasAccount: CcdAmount.Type;
     // The new balance of the GAS account.
-    newGasAccount: Amount;
+    newGasAccount: CcdAmount.Type;
     // The amount awarded to the baker.
-    bakerReward: Amount;
+    bakerReward: CcdAmount.Type;
     // The amount awarded to the foundation.
-    foundationCharge: Amount;
+    foundationCharge: CcdAmount.Type;
     // The baker of the block, who receives the award.
-    baker: Base58String;
+    baker: AccountAddress.Type;
     // The foundation account.
-    foundationAccount: Base58String;
+    foundationAccount: AccountAddress.Type;
 }
 
 export interface BlockSpecialEventPaydayFoundationReward {
     tag: 'paydayFoundationReward';
     // The account that got rewarded.
-    foundationAccount: Base58String;
+    foundationAccount: AccountAddress.Type;
     // The transaction fee reward at payday to the account.
-    developmentCharge: Amount;
+    developmentCharge: CcdAmount.Type;
 }
 
 export interface BlockSpecialEventPaydayAccountReward {
     tag: 'paydayAccountReward';
     // The account that got rewarded.
-    account: Base58String;
+    account: AccountAddress.Type;
     // The transaction fee reward at payday to the account.
-    transactionFees: Amount;
+    transactionFees: CcdAmount.Type;
     // The baking reward at payday to the account.
-    bakerReward: Amount;
+    bakerReward: CcdAmount.Type;
     // The finalization reward at payday to the account.
-    finalizationReward: Amount;
+    finalizationReward: CcdAmount.Type;
 }
 
 export interface BlockSpecialEventBlockAccrueReward {
     tag: 'blockAccrueReward';
     // The total fees paid for transactions in the block.
-    transactionFees: Amount;
+    transactionFees: CcdAmount.Type;
     // The old balance of the GAS account.
-    oldGasAccount: Amount;
+    oldGasAccount: CcdAmount.Type;
     // The new balance of the GAS account.
-    newGasAccount: Amount;
+    newGasAccount: CcdAmount.Type;
     // The amount awarded to the baker.
-    bakerReward: Amount;
+    bakerReward: CcdAmount.Type;
     // The amount awarded to the passive delegators.
-    passiveReward: Amount;
+    passiveReward: CcdAmount.Type;
     // The amount awarded to the foundation.
-    foundationCharge: Amount;
+    foundationCharge: CcdAmount.Type;
     // The baker of the block, who will receive the award.
     baker: BakerId;
 }
@@ -99,33 +101,33 @@ export interface BlockSpecialEventPaydayPoolReward {
     // The pool owner (passive delegators when not present).
     poolOwner?: BakerId;
     // Accrued transaction fees for pool.
-    transactionFees: Amount;
+    transactionFees: CcdAmount.Type;
     // Accrued baking rewards for pool.
-    bakerReward: Amount;
+    bakerReward: CcdAmount.Type;
     // Accrued finalization rewards for pool.
-    finalizationReward: Amount;
+    finalizationReward: CcdAmount.Type;
 }
 
 export interface BlockSpecialEventAccountAmount {
     // The key type
-    account: Base58String;
+    account: AccountAddress.Type;
     // The value type
-    amount: Amount;
+    amount: CcdAmount.Type;
 }
 
 /**
- * Gets a list of {@link Base58String} account addresses affected the {@link BlockSpecialEvent}.
+ * Gets a list of {@link AccountAddress.Type} account addresses affected the {@link BlockSpecialEvent}.
  *
  * @param {BlockSpecialEvent} event - The block special event to check.
  *
- * @returns {Base58String[]} List of account addresses affected by the event.
+ * @returns {AccountAddress.Type[]} List of account addresses affected by the event.
  */
 export function specialEventAffectedAccounts(
     event: Exclude<
         BlockSpecialEvent,
         BlockSpecialEventBlockAccrueReward | BlockSpecialEventPaydayPoolReward
     >
-): Base58String[];
+): AccountAddress.Type[];
 export function specialEventAffectedAccounts(
     event:
         | BlockSpecialEventBlockAccrueReward
@@ -133,10 +135,10 @@ export function specialEventAffectedAccounts(
 ): never[];
 export function specialEventAffectedAccounts(
     event: BlockSpecialEvent
-): Base58String[];
+): AccountAddress.Type[];
 export function specialEventAffectedAccounts(
     event: BlockSpecialEvent
-): Base58String[] {
+): AccountAddress.Type[] {
     switch (event.tag) {
         case 'bakingRewards':
             return event.bakingRewards.map((br) => br.account);
@@ -148,7 +150,7 @@ export function specialEventAffectedAccounts(
         case 'paydayAccountReward':
             return [event.account];
         case 'blockReward': {
-            if (event.baker === event.foundationAccount) {
+            if (AccountAddress.equals(event.baker, event.foundationAccount)) {
                 return [event.baker];
             }
             return [event.baker, event.foundationAccount];