+++ packages/sdk/test/ci/types/blockSpecialEvents.test.ts
@@ -8,99 +8,123 @@ import {
     BlockSpecialEventPaydayAccountReward,
     BlockSpecialEventBlockReward,
     specialEventAffectedAccounts,
+    CcdAmount,
+    AccountAddress,
 } from '../../../src/index.js';
 
 const bakingRewards: BlockSpecialEventBakingRewards = {
     tag: 'bakingRewards',
-    remainder: 0n,
+    remainder: CcdAmount.zero(),
     bakingRewards: [
         {
-            account: '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-            amount: 400000n,
+            account: AccountAddress.fromBase58(
+                '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe'
+            ),
+            amount: CcdAmount.fromMicroCcd(400000),
         },
         {
-            account: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-            amount: 400000n,
+            account: AccountAddress.fromBase58(
+                '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+            ),
+            amount: CcdAmount.fromMicroCcd(400000),
         },
     ],
 };
 
 const finalizationRewards: BlockSpecialEventFinalizationRewards = {
     tag: 'finalizationRewards',
-    remainder: 0n,
+    remainder: CcdAmount.zero(),
     finalizationRewards: [
         {
-            account: '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-            amount: 400000n,
+            account: AccountAddress.fromBase58(
+                '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe'
+            ),
+            amount: CcdAmount.fromMicroCcd(400000),
         },
         {
-            account: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-            amount: 400000n,
+            account: AccountAddress.fromBase58(
+                '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+            ),
+            amount: CcdAmount.fromMicroCcd(400000),
         },
     ],
 };
 
 const foundationReward: BlockSpecialEventPaydayFoundationReward = {
     tag: 'paydayFoundationReward',
-    developmentCharge: 123n,
-    foundationAccount: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+    developmentCharge: CcdAmount.fromMicroCcd(123),
+    foundationAccount: AccountAddress.fromBase58(
+        '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+    ),
 };
 
 const mint: BlockSpecialEventMint = {
     tag: 'mint',
-    mintBakingReward: 0n,
-    mintFinalizationReward: 0n,
-    mintPlatformDevelopmentCharge: 0n,
-    foundationAccount: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+    mintBakingReward: CcdAmount.zero(),
+    mintFinalizationReward: CcdAmount.zero(),
+    mintPlatformDevelopmentCharge: CcdAmount.zero(),
+    foundationAccount: AccountAddress.fromBase58(
+        '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+    ),
 };
 
 const paydayAccountReward: BlockSpecialEventPaydayAccountReward = {
     tag: 'paydayAccountReward',
-    account: '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-    transactionFees: 123n,
-    bakerReward: 123n,
-    finalizationReward: 123n,
+    account: AccountAddress.fromBase58(
+        '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe'
+    ),
+    transactionFees: CcdAmount.fromMicroCcd(123),
+    bakerReward: CcdAmount.fromMicroCcd(123),
+    finalizationReward: CcdAmount.fromMicroCcd(123),
 };
 
 const foundationBlockReward: BlockSpecialEventBlockReward = {
     tag: 'blockReward',
-    transactionFees: 1231241n,
-    bakerReward: 12314n,
-    foundationCharge: 12n,
-    newGasAccount: 1n,
-    oldGasAccount: 0n,
-    baker: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-    foundationAccount: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+    transactionFees: CcdAmount.fromMicroCcd(1231241),
+    bakerReward: CcdAmount.fromMicroCcd(12314),
+    foundationCharge: CcdAmount.fromMicroCcd(12),
+    newGasAccount: CcdAmount.fromMicroCcd(1),
+    oldGasAccount: CcdAmount.zero(),
+    baker: AccountAddress.fromBase58(
+        '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+    ),
+    foundationAccount: AccountAddress.fromBase58(
+        '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+    ),
 };
 
 const blockReward: BlockSpecialEventBlockReward = {
     tag: 'blockReward',
-    transactionFees: 1231241n,
-    bakerReward: 12314n,
-    foundationCharge: 12n,
-    newGasAccount: 1n,
-    oldGasAccount: 0n,
-    baker: '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-    foundationAccount: '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+    transactionFees: CcdAmount.fromMicroCcd(1231241),
+    bakerReward: CcdAmount.fromMicroCcd(12314),
+    foundationCharge: CcdAmount.fromMicroCcd(12),
+    newGasAccount: CcdAmount.fromMicroCcd(1),
+    oldGasAccount: CcdAmount.zero(),
+    baker: AccountAddress.fromBase58(
+        '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe'
+    ),
+    foundationAccount: AccountAddress.fromBase58(
+        '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'
+    ),
 };
 
 const paydayPoolReward: BlockSpecialEventPaydayPoolReward = {
     tag: 'paydayPoolReward',
     poolOwner: 123n,
-    finalizationReward: 123n,
-    bakerReward: 12314n,
-    transactionFees: 1231241n,
+    finalizationReward: CcdAmount.fromMicroCcd(123),
+    bakerReward: CcdAmount.fromMicroCcd(12314),
+    transactionFees: CcdAmount.fromMicroCcd(1231241),
 };
 
 const accrueReward: BlockSpecialEventBlockAccrueReward = {
     tag: 'blockAccrueReward',
-    transactionFees: 1231241n,
-    bakerReward: 12314n,
+    transactionFees: CcdAmount.fromMicroCcd(1231241),
+    bakerReward: CcdAmount.fromMicroCcd(12314),
     baker: 0n,
-    foundationCharge: 12n,
-    newGasAccount: 1n,
-    oldGasAccount: 0n,
-    passiveReward: 123n,
+    foundationCharge: CcdAmount.fromMicroCcd(12),
+    newGasAccount: CcdAmount.fromMicroCcd(1),
+    oldGasAccount: CcdAmount.zero(),
+    passiveReward: CcdAmount.fromMicroCcd(123),
 };
 
 describe('specialEventAffectedAccounts', () => {
@@ -114,41 +138,55 @@ describe('specialEventAffectedAccounts', () => {
 
     test('Returns correct list of accounts for events with account payouts', () => {
         let accounts = specialEventAffectedAccounts(bakingRewards);
-        expect(accounts).toEqual([
-            '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-            '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-        ]);
+        expect(accounts).toEqual(
+            [
+                '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
+                '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+            ].map(AccountAddress.fromBase58)
+        );
 
         accounts = specialEventAffectedAccounts(finalizationRewards);
-        expect(accounts).toEqual([
-            '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-            '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-        ]);
+        expect(accounts).toEqual(
+            [
+                '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
+                '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+            ].map(AccountAddress.fromBase58)
+        );
 
         accounts = specialEventAffectedAccounts(foundationReward);
-        expect(accounts).toEqual([
-            '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-        ]);
+        expect(accounts).toEqual(
+            ['3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'].map(
+                AccountAddress.fromBase58
+            )
+        );
 
         accounts = specialEventAffectedAccounts(mint);
-        expect(accounts).toEqual([
-            '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-        ]);
+        expect(accounts).toEqual(
+            ['3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'].map(
+                AccountAddress.fromBase58
+            )
+        );
 
         accounts = specialEventAffectedAccounts(paydayAccountReward);
-        expect(accounts).toEqual([
-            '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-        ]);
+        expect(accounts).toEqual(
+            ['4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe'].map(
+                AccountAddress.fromBase58
+            )
+        );
 
         accounts = specialEventAffectedAccounts(foundationBlockReward);
-        expect(accounts).toEqual([
-            '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-        ]);
+        expect(accounts).toEqual(
+            ['3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV'].map(
+                AccountAddress.fromBase58
+            )
+        );
 
         accounts = specialEventAffectedAccounts(blockReward);
-        expect(accounts).toEqual([
-            '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
-            '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
-        ]);
+        expect(accounts).toEqual(
+            [
+                '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe',
+                '3v1JUB1R1JLFtcKvHqD9QFqe2NXeBF53tp69FLPHYipTjNgLrV',
+            ].map(AccountAddress.fromBase58)
+        );
     });
 });